+++ lib/pages/edit_verse/edit_verse_manager.dart
@@ -1,0 +1,39 @@
+import 'package:flutter/foundation.dart';
+import 'package:memorize_scripture/common/verse.dart';
+import 'package:memorize_scripture/service_locator.dart';
+import 'package:memorize_scripture/services/data_repository.dart';
+
+class EditVersePageManager {
+  final verseNotifier = ValueNotifier<Verse?>(null);
+  final dataRepo = getIt<DataRepository>();
+
+  late String _collectionId;
+  late String _verseId;
+
+  Future<void> init({
+    required String collectionId,
+    required String verseId,
+  }) async {
+    _collectionId = collectionId;
+    _verseId = verseId;
+
+    verseNotifier.value = await dataRepo.fetchVerse(
+      collectionId: collectionId,
+      verseId: verseId,
+    );
+  }
+
+  Future<void> saveVerse({
+    required String prompt,
+    required String answer,
+  }) async {
+    await dataRepo.updateVerse(
+      _collectionId,
+      Verse(
+        id: _verseId,
+        prompt: prompt,
+        answer: answer,
+      ),
+    );
+  }
+}