+++ lib/services/sample_verses.dart
@@ -2,214 +2,214 @@ import 'package:memorize_scripture/common/verse.dart';
 
 final starterVersesWeb = [
   Verse(
-    translation: 'WEB',
-    prompt: 'John 3:16',
+    id: 's01',
+    prompt: 'John 3:16 (WEB)',
     answer:
         'For God so loved the world, that he gave his one and only Son, that whoever believes in him should not perish, but have eternal life.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Matthew 28:19-20',
+    id: 's02',
+    prompt: 'Matthew 28:19-20 (WEB)',
     answer:
         'Go and make disciples of all nations, baptizing them in the name of the Father and of the Son and of the Holy Spirit, teaching them to observe all things that I commanded you. Behold, I am with you always, even to the end of the age.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Philippians 4:13',
+    id: 's03',
+    prompt: 'Philippians 4:13 (WEB)',
     answer: 'I can do all things through Christ, who strengthens me.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Galatians 2:20',
+    id: 's04',
+    prompt: 'Galatians 2:20 (WEB)',
     answer:
         'I have been crucified with Christ, and it is no longer I who live, but Christ lives in me. That life which I now live in the flesh, I live by faith in the Son of God, who loved me, and gave himself up for me.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Proverbs 3:5-6',
+    id: 's05',
+    prompt: 'Proverbs 3:5-6 (WEB)',
     answer:
         'Trust in Yahweh with all your heart, and don’t lean on your own understanding. In all your ways acknowledge him, and he will make your paths straight.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Romans 8:28',
+    id: 's06',
+    prompt: 'Romans 8:28 (WEB)',
     answer:
         'We know that all things work together for good for those who love God, for those who are called according to his purpose.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: '1 Thessalonians 5:16-18',
+    id: 's07',
+    prompt: '1 Thessalonians 5:16-18 (WEB)',
     answer:
         'Always rejoice. Pray without ceasing. In everything give thanks, for this is the will of God in Christ Jesus toward you.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: '2 Corinthians 5:17',
+    id: 's08',
+    prompt: '2 Corinthians 5:17 (WEB)',
     answer:
         'Therefore if anyone is in Christ, he is a new creation. The old things have passed away. Behold, all things have become new. ',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: '2 Timothy 3:16-17',
+    id: 's09',
+    prompt: '2 Timothy 3:16-17 (WEB)',
     answer:
         'Every Scripture is God-breathed and profitable for teaching, for reproof, for correction, and for instruction in righteousness, that each person who belongs to God may be complete, thoroughly equipped for every good work.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Ephesians 2:8-9',
+    id: 's10',
+    prompt: 'Ephesians 2:8-9 (WEB)',
     answer:
         'for by grace you have been saved through faith, and that not of yourselves; it is the gift of God, not of works, that no one would boast.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Ephesians 4:32',
+    id: 's11',
+    prompt: 'Ephesians 4:32 (WEB)',
     answer:
         'And be kind to one another, tender hearted, forgiving each other, just as God also in Christ forgave you.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'John 14:6',
+    id: 's12',
+    prompt: 'John 14:6 (WEB)',
     answer:
         'I am the way, the truth, and the life. No one comes to the Father, except through me.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Matthew 5:16',
+    id: 's13',
+    prompt: 'Matthew 5:16 (WEB)',
     answer:
         'Even so, let your light shine before men, that they may see your good works and glorify your Father who is in heaven.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Matthew 6:33',
+    id: 's14',
+    prompt: 'Matthew 6:33 (WEB)',
     answer:
         'But seek first God’s Kingdom and his righteousness; and all these things will be given to you as well.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Romans 3:23',
+    id: 's15',
+    prompt: 'Romans 3:23 (WEB)',
     answer: 'for all have sinned, and fall short of the glory of God;',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Romans 5:8',
+    id: 's16',
+    prompt: 'Romans 5:8 (WEB)',
     answer:
         'But God commends his own love toward us, in that while we were yet sinners, Christ died for us.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Romans 6:23',
+    id: 's17',
+    prompt: 'Romans 6:23 (WEB)',
     answer:
         'For the wages of sin is death, but the free gift of God is eternal life in Christ Jesus our Lord.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: '1 John 1:9',
+    id: 's18',
+    prompt: '1 John 1:9 (WEB)',
     answer:
         'If we confess our sins, he is faithful and righteous to forgive us the sins, and to cleanse us from all unrighteousness.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: '1 Peter 5:6-7',
+    id: 's19',
+    prompt: '1 Peter 5:6-7 (WEB)',
     answer:
         'Humble yourselves therefore under the mighty hand of God, that he may exalt you in due time, casting all your worries on him, because he cares for you.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: '2 Corinthians 9:7',
+    id: 's20',
+    prompt: '2 Corinthians 9:7 (WEB)',
     answer:
         'Let each man give according as he has determined in his heart, not grudgingly or under compulsion, for God loves a cheerful giver.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Galatians 5:22-23',
+    id: 's21',
+    prompt: 'Galatians 5:22-23 (WEB)',
     answer:
         'But the fruit of the Spirit is love, joy, peace, patience, kindness, goodness, faith, gentleness, and self-control. Against such things there is no law.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Hebrews 11:1',
+    id: 's22',
+    prompt: 'Hebrews 11:1 (WEB)',
     answer:
         'Now faith is assurance of things hoped for, proof of things not seen.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Hebrews 4:16',
+    id: 's23',
+    prompt: 'Hebrews 4:16 (WEB)',
     answer:
         'Let’s therefore draw near with boldness to the throne of grace, that we may receive mercy and may find grace for help in time of need.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Jeremiah 29:11',
+    id: 's24',
+    prompt: 'Jeremiah 29:11 (WEB)',
     answer:
         'For I know the thoughts that I think toward you,” says Yahweh, “thoughts of peace, and not of evil, to give you hope and a future.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Joshua 1:9',
+    id: 's25',
+    prompt: 'Joshua 1:9 (WEB)',
     answer:
         'Haven’t I commanded you? Be strong and courageous. Don’t be afraid. Don’t be dismayed, for Yahweh your God is with you wherever you go.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Lamentations 3:22-23',
+    id: 's26',
+    prompt: 'Lamentations 3:22-23 (WEB)',
     answer:
         'It is because of Yahweh’s loving kindnesses that we are not consumed, because his compassion doesn’t fail. They are new every morning. Great is your faithfulness.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Luke 9:23',
+    id: 's27',
+    prompt: 'Luke 9:23 (WEB)',
     answer:
         'If anyone desires to come after me, let him deny himself, take up his cross, and follow me.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Matthew 11:28',
+    id: 's28',
+    prompt: 'Matthew 11:28 (WEB)',
     answer:
         'Come to me, all you who labor and are heavily burdened, and I will give you rest.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Matthew 4:19',
+    id: 's29',
+    prompt: 'Matthew 4:19 (WEB)',
     answer:
         'He said to them, “Come after me, and I will make you fishers for men.”',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Matthew 6:34',
+    id: 's30',
+    prompt: 'Matthew 6:34 (WEB)',
     answer:
         'Therefore don’t be anxious for tomorrow, for tomorrow will be anxious for itself. Each day’s own evil is sufficient.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Philippians 4:4',
+    id: 's31',
+    prompt: 'Philippians 4:4 (WEB)',
     answer: 'Rejoice in the Lord always! Again I will say, “Rejoice!”',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Philippians 4:6-7',
+    id: 's32',
+    prompt: 'Philippians 4:6-7 (WEB)',
     answer:
         'In nothing be anxious, but in everything, by prayer and petition with thanksgiving, let your requests be made known to God. 7 And the peace of God, which surpasses all understanding, will guard your hearts and your thoughts in Christ Jesus.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Psalm 46:1',
+    id: 's33',
+    prompt: 'Psalm 46:1 (WEB)',
     answer: 'God is our refuge and strength, a very present help in trouble.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Psalm 46:10',
+    id: 's34',
+    prompt: 'Psalm 46:10 (WEB)',
     answer:
         'Be still, and know that I am God. I will be exalted among the nations. I will be exalted in the earth.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Revelation 3:20',
+    id: 's35',
+    prompt: 'Revelation 3:20 (WEB)',
     answer:
         'Behold, I stand at the door and knock. If anyone hears my voice and opens the door, then I will come in to him, and will dine with him, and he with me.',
   ),
   Verse(
-    translation: 'WEB',
-    prompt: 'Romans 8:32',
+    id: 's36',
+    prompt: 'Romans 8:32 (WEB)',
     answer:
         'He who didn’t spare his own Son, but delivered him up for us all, how would he not also with him freely give us all things?',
   ),